% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_maxent.R
\name{binarize_pred}
\alias{binarize_pred}
\title{Binarize a prediction raster}
\usage{
binarize_pred(sdm, threshold)
}
\arguments{
\item{obs}{data frame, containing the coordinates to reproject}

\item{predictors, }{raster}

\item{lon}{string, name of the longitude column (same projection as predictor raster)}

\item{lat}{string, name of the latitude column (same projection as predictor raster)}

\item{proj}{character, initial projection of the xy coordinates}
}
\value{
spatial points
}
\description{
Binarize a prediction raster
}
