% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_background.R
\name{create_background}
\alias{create_background}
\title{Create background points}
\usage{
create_background(
  predictors,
  mask = NULL,
  method = "random",
  n = 10000,
  obs = NULL,
  density_bias = NULL,
  width_buffer = NULL,
  species = NULL
)
}
\arguments{
\item{predictors}{spat raster, containing the predictor variables}

\item{mask, }{spat vector, mask to apply to the predictors.}

\item{method}{string, random, inclusion_buffer, thickening or biased.}

\item{n}{integer, number of background points to select.}

\item{obs}{data.frame, containing the observations. Used with method == "thickening" or "inclusion_buffer"}

\item{width_buffer}{int, buffer width around observations. Used with method ==  "inclusion_buffer"}

\item{species}{string, species name.}
}
\value{
spatial points
}
\description{
Create background points
}
