% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_study_extent.R
\name{create_study_extent}
\alias{create_study_extent}
\title{Create study extent}
\usage{
create_study_extent(
  obs,
  lon = "lon",
  lat = "lat",
  proj = NULL,
  method = "box",
  width_buffer = NULL,
  mask = NULL,
  shapefile_path = NULL
)
}
\arguments{
\item{obs}{data frame, containing the observations}

\item{lon}{string, name of the longitude column}

\item{lat}{string, name of the latitude column}

\item{proj}{character, initial projection of the xy coordinates}

\item{method}{string, "box", "mcp", "buffer" or "user_shapefile"}

\item{width_buffer}{int, buffer width around observations, box or mcp.}

\item{mask, }{spat vector, mask to apply to the study area.}

\item{shapefile_path, }{string, path to the user shapefile.}
}
\value{
spatial points
}
\description{
Create study extent
}
