% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_maxent.R
\name{find_threshold}
\alias{find_threshold}
\title{Find threshold to binarize an sdm output}
\usage{
find_threshold(sdm, occs, bg, type = "mtp")
}
\arguments{
\item{sdm}{raster, sdm prediction}

\item{occs}{dataframe, presence points}

\item{bg}{dataframe, background points}

\item{type}{string, method to choose the threshold, one of "mtp", "kappa", "spec_sens", "no_omission", "prevalence", "equal_sens_spec", "sensitivity"}
}
\value{
sa float
}
\description{
Find threshold to binarize an sdm output
}
