% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_observations.R
\name{load_observations}
\alias{load_observations}
\title{Load observations from GBIF or ATLAS databases}
\usage{
load_observations(
  species,
  data_source,
  year_start,
  year_end = NULL,
  extent_wkt = NULL,
  extent_shp = NULL,
  proj_shp = NULL,
  xmin = NA,
  ymin = NA,
  xmax = NA,
  ymax = NA,
  bbox = NULL,
  limit = 1000
)
}
\arguments{
\item{species, }{a character, scientific name of the species.}

\item{data_source, }{an integer, "gbif" or "atlas", source to use}

\item{year_start, }{an integer, starting year of the period}

\item{year_end, }{an integer, end year of the period}

\item{extent_wkt, }{a string in WKT format, used to subset the obs spatially (only used with gbif for now)}

\item{extent_shp, }{sf polygone, used to subset the obs spatially (only used with gbif for now)}

\item{proj_shp, }{extent_shp projection system, if not included in extent_shp.}

\item{xmin, }{a float, xmin coords of extent box (MUST BE IN WGS84), used to subset the obs spatially (only used with gbif for now)}

\item{ymin, }{a float, ymin coords of extent box (MUST BE IN WGS84), used to subset the obs spatially (only used with gbif for now)#' @param aggregation, a character, aggregation method as string, defining how to deal with pixels containing data from multiple images, can be "min", "max", "mean", "median", or "first"}

\item{xmax, }{a float, xmax coords of extent box (MUST BE IN WGS84), used to subset the obs spatially (only used with gbif for now)#' @return a raster stack of variables not intercorrelated}

\item{ymax, }{a float, ymax coords of extent box (MUST BE IN WGS84), used to subset the obs spatially (only used with gbif for now)}

\item{bbox, }{a vector of float, format: c(xmin, ymin, xmax, ymax) (MUST BE IN WGS84), used to subset the obs spatially (only used with gbif for now)}

\item{limit, }{an integer, limit param for rgbif::occ_data function}
}
\value{
a data.frame
}
\description{
Load observations from GBIF or ATLAS databases
}
