% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_predictors.R
\name{loadPredictors}
\alias{loadPredictors}
\alias{load_predictors}
\title{Remove collinear variables}
\usage{
load_predictors(
  layersDir,
  subsetLayers = NULL,
  removeCollinear = T,
  method = "vif.cor",
  method.cor.vif = NULL,
  proj = NULL,
  mask = NULL,
  sample = TRUE,
  nb.points = 5000,
  cutoff.cor = 0.7,
  cutoff.vif = 3,
  export = TRUE,
  loadNonCollinear = F,
  nonCollinearDir = NULL
)
}
\arguments{
\item{layersDir, }{string, path to the folder containing the initial raster layers}

\item{subsetLayers, }{a vector, containing the name of layers to select. If NULL, all layers in dir.pred selected by default.}

\item{removeCollinear, }{boolean. If TRUE, an analysis of collinearity is performed.}

\item{method, }{The correlation method to be used:"vif.cor", "vif.step", "pearson", "spearman"
or "kendall". "vif.cor" and "vif.step" use the Variance Inflation factor and the pearson correlation, more details
here https://www.rdocumentation.org/packages/usdm/versions/1.1-18/topics/vif. If your variables are skewed or have outliers
(e.g. when working with precipitation variables) you should favour the Spearman or Kendall methods.}

\item{method.cor.vif, }{the correlation method to be used with "vif.cor" method. "pearson", "spearman"
or "kendall".}

\item{proj, }{a string , proj if raster layers has to be reprojected}

\item{mask, }{a polygon to crop the raster}

\item{sample, }{boolean value. If TRUE, sample a number of points equal to nb.points before evaluating collinearity}

\item{nb.points, }{a numeric value. Only used if sample.points = TRUE. The number of sampled points from the raster.}

\item{cutoff.cor, }{a numeric value corresponding to the maximum threshold of linear correlation (for "vif.cor", "pearson", "spearman").}

\item{cutoff.vif, }{a numeric value corresponding to the maximum threshold of VIF (only used for method "vif.step").}

\item{export, }{boolean value. If TRUE, the list of selected variables and the correlation matrix will be saved in nonCollinearDir.}

\item{loadNonCollinear, }{boolean, if TRUE, a list of non-collinear layers is used to subset the raster stack (loaded from nonCollinearDir)}

\item{nonCollinearDir, }{string, path to the folder export or impot the list of non-collinear layers (if loadNonCollinear = T)}
}
\value{
a raster stack of variables not intercorrelated
}
\description{
Remove collinear variables
}
