% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_maxent.R
\name{predict_maxent}
\alias{predict_maxent}
\title{Predict with a tuned maxEnt model}
\usage{
predict_maxent(mod, param, predictors, type = "cloglog", mask = NULL)
}
\arguments{
\item{mod}{ENMevaluation object}

\item{param}{string, maxent select parameters e.g. "fc.L_rm.1"}

\item{predictors}{spat Raster, environmental predictor variables.}

\item{type}{string, type of output "cloglog" or "raw"}

\item{mask}{spatVector, a mask}
}
\value{
raster
}
\description{
Predict with a tuned maxEnt model
}
