% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeCollinearity.R
\name{remove_collinearity}
\alias{remove_collinearity}
\title{Remove collinear variables}
\usage{
remove_collinearity(
  predictors,
  method = "vif.cor",
  method.cor.vif = NULL,
  sample = FALSE,
  nb.points = 5000,
  cutoff.cor = 0.7,
  cutoff.vif = 10,
  export = FALSE,
  title_export = "",
  path = NULL
)
}
\arguments{
\item{predictors, }{a raster, either from raster or terra format}

\item{method, }{The correlation method to be used:"vif.cor", "vif.step", "pearson", "spearman"
or "kendall". "vif.cor" and "vif.step" use the Variance Inflation factor and the pearson correlation, more details
here https://www.rdocumentation.org/packages/usdm/versions/1.1-18/topics/vif. If your variables are skewed or have outliers
(e.g. when working with precipitation variables) you should favour the Spearman or Kendall methods.}

\item{method.cor.vif, }{the correlation method to be used with "vif.cor" method. "pearson", "spearman"
or "kendall".}

\item{sample, }{boolean value. If TRUE, sample a number of points equal to nb.points before evaluating collinearity}

\item{nb.points, }{a numeric value. Only used if sample.points = TRUE. The number of sampled points from the raster.}

\item{cutoff.cor, }{a numeric value corresponding to the maximum threshold of linear correlation (for "vif.cor", "pearson", "spearman").}

\item{cutoff.vif, }{a numeric value corresponding to the maximum threshold of VIF (only used for method "vif.step").}

\item{export, }{boolean value. If TRUE, the list of selected variables and the correlation matrix will be saved.}

\item{title_export, }{a string. If export is TRUE, title of the correlation matrix.}

\item{path, }{a string. If export is TRUE, path to save the corrrelation matrix and list of uncorrelated variables.}

\item{mask, }{a}
}
\value{
a raster stack of variables not intercorrelated
}
\description{
Remove collinear variables
}
