% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_maxent.R
\name{response_plot}
\alias{response_plot}
\title{Create response plot}
\usage{
response_plot(mod, param, type = "cloglog")
}
\arguments{
\item{mod}{ENMevaluation object}

\item{param}{string, maxent select parameters e.g. "fc.L_rm.1"}

\item{type}{string, type of output "cloglog" or "raw"}
}
\value{
a plot
}
\description{
Create response plot
}
