% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_maxent.R
\name{run_maxent}
\alias{run_maxent}
\title{Run maxEnt model}
\usage{
run_maxent(
  presence.bg,
  with_raster = F,
  algorithm = "maxnet",
  factors = NULL,
  predictors = NULL,
  covars = NULL,
  partition_type = "crossvalidation",
  nfolds = 5,
  orientation_block = "lat_lon",
  fc = "L",
  rm = 1,
  parallel = T,
  updateProgress = T,
  parallelType = "doParallel"
)
}
\arguments{
\item{presence.bg}{data frame, containing presence/background, output of setup_sdm_data}

\item{with_raster}{boolean, if F presence.bg must contain the predictors values.}

\item{algorithm}{string, "maxnet" or "maxent.jar"}

\item{factors}{character vector, name or names of categorical environmental variables.}

\item{predictors}{spatRaster, environmental predictor variables. Used if with_raster = T.}

\item{covars}{character vector, if with_raster = F, list of predictors variables (must match columns names in presence.bg)}

\item{partition_type}{string, name of partitioning technique: "randomkfold","jackknife","block", "checkerboard1", "checkerboard2", and "none"}

\item{nfolds}{int, number of folds (for "randomkfold" method)}

\item{orientation_block}{string, one of "lat_lon" (default), "lon_lat", "lat_lat", or "lon_lon"}

\item{fc}{vector of strings, feature classes to test, e.g. c("L", "LQ")}

\item{rm}{vector of float, regularisation multiplier values to test, e.g. c(1,2,3)}

\item{parallel}{boolean, if TRUE, run with parallel processing.}

\item{updateProgress}{boolean, if TRUE, use shiny progress bar. This is only for use in shiny apps.}

\item{parallelType}{character, either "doParallel" or "doSNOW" (default: "doSNOW") .}
}
\value{
spatial points
}
\description{
Run maxEnt model
}
