% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_maxent.R
\name{select_param}
\alias{select_param}
\title{Select best maxEnt parameters}
\usage{
select_param(res, method = "AIC", auc_min = 0, list = T)
}
\arguments{
\item{res}{data frame, containing result slot of an ENMevaluation object}

\item{method}{string, "AIC" or "p10"}

\item{auc_min}{float, minimum value of the AUC}

\item{list}{boolean, if F the param are returned in the form "fc.L_rm.1", if T as a named list.}
}
\value{
list or string
}
\description{
Select best maxEnt parameters
}
