% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_sdm_data.R
\name{setup_sdm_data}
\alias{setup_sdm_data}
\title{Create a datasets to feed any sdm algorithm}
\usage{
setup_sdm_data(
  presence,
  background,
  predictors,
  partition_type = "none",
  runs_n = 2,
  boot_proportion = 0.7,
  cv_partitions = NULL,
  seed = NULL
)
}
\arguments{
\item{presence}{data frame, containing the presence}

\item{background}{data frame, containing the absence or background points}

\item{predictors}{spat raster, predictors}

\item{partition_type}{string, partition method (not yet used) LET TO "none"}

\item{runs_n}{int, number of runs (not yet used)}

\item{boot_proportion}{int, number of runs (not yet used)}

\item{cv_partitions}{int, (not yet used)}

\item{seed}{(not yet used)}
}
\value{
a data frame
}
\description{
Create a datasets to feed any sdm algorithm
}
